+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.3.4 (2020-06-25, https://github.com/go-reform/reform/milestones/v1.3.4)
+
+* Make reform generator work with Go 1.15.
+* Replace syreclabs.com/go/faker with github.com/brianvoe/gofakeit.
+
 ## v1.3.3 (2018-12-11, https://github.com/go-reform/reform/milestones/v1.3.3)
 
 * Fix tests for Go 1.12.